1.21

+++ yu-ai-agent/yu-ai-agent-master/src/main/java/com/yupi/yuaiagent/domin/entity/Borehole.java
@@ -38,7 +38,7 @@ public class Borehole {
     /**
      * 总深度
      */
-    private BigDecimal totalDepth;
+    private Double totalDepth;
 
     /**
      * 钻探目的
